Adjust API usage to use CORS
This removes API_PREFIX_PATH constant in favor of supporting requests
via CORS from the web app since the web app is no longer using a proxy
to this app.

This tweaks the headers to allow access to the web app, which uses
localhost:4200 for development and an ENV for production. localhost:4200
is the default Ember CLI port, so it seems like a sane default.